+++ packages/imperative/src/config/cmd/secure/secure.handler.ts
@@ -9,12 +9,24 @@
 *
 */
 
-import { ICommandHandler, IHandlerParameters } from "../../../../../cmd";
+import * as lodash from "lodash";
+import { ICommandArguments, ICommandHandler, IHandlerParameters } from "../../../../../cmd";
+import { ICommandHandlerRequire } from "../../../../../cmd/src/doc/handler/ICommandHandlerRequire";
+import { ICommandProfileAuthConfig } from "../../../../../cmd/src/doc/profiles/definition/ICommandProfileAuthConfig";
+import { Config } from "../../../../../config";
 import { secureSaveError } from "../../../../../config/src/ConfigUtils";
+import { ImperativeError } from "../../../../../error";
+import { Logger } from "../../../../../logger";
+import { ConnectionPropsForSessCfg, ISession, Session } from "../../../../../rest";
 import { CredentialManagerFactory } from "../../../../../security";
 import { ImperativeConfig } from "../../../../../utilities";
+import { BaseAuthHandler } from "../../../auth/handlers/BaseAuthHandler";
 
 export default class SecureHandler implements ICommandHandler {
+    /**
+     * The parameters object passed to the command handler.
+     */
+    private params: IHandlerParameters;
 
     /**
      * Process the command and input.
@@ -24,6 +36,7 @@ export default class SecureHandler implements ICommandHandler {
      * @throws {ImperativeError}
      */
     public async process(params: IHandlerParameters): Promise<void> {
+        this.params = params;
 
         // Setup the credential vault API for the config
         if (!CredentialManagerFactory.initialized) {
@@ -41,7 +54,13 @@ export default class SecureHandler implements ICommandHandler {
         }
 
         // Prompt for values designated as secure
+        let authTokenProp: string;
         for (const propName of secureProps) {
+            if (authTokenProp == null && propName.endsWith(".tokenValue")) {
+                authTokenProp = propName;
+                continue;
+            }
+
             const propValue = await params.response.console.prompt(`Please enter ${propName}: `, {hideText: true});
 
             // Save the value in the config securely
@@ -50,7 +69,69 @@ export default class SecureHandler implements ICommandHandler {
             }
         }
 
+        if (authTokenProp != null) {
+            const propValue = await this.handlePromptForAuthToken(config, authTokenProp) ||
+                await params.response.console.prompt(`Please enter ${authTokenProp}: `, {hideText: true});
+
+            // Save the value in the config securely
+            if (propValue) {
+                config.set(authTokenProp, propValue, { secure: true });
+            }
+        }
+
         // Write the config layer
         await config.save(false);
+    }
+
+    /**
+     * Checks if authentication service is associated with an auth token
+     * property. If an auth service is found, we log in to it obtain the token
+     * instead of prompting for the value.
+     * @param config Team config class from which the property was loaded
+     * @param propPath JSON property path of the auth token
+     * @returns Token value, or undefined if none was obtained
+     */
+    private async handlePromptForAuthToken(config: Config, propPath: string): Promise<string | undefined> {
+        const profilePath = propPath.slice(0, propPath.indexOf(".properties"));
+        const profileType = lodash.get(config.properties, `${profilePath}.type`);
+        const profile = config.api.profiles.get(profilePath.replace(/profiles\./g, ""));
+        if (profileType == null || profile.tokenType == null) {
+            return;  // Can't find auth service if profile type or token type are unknown
+        }
+
+        const authConfigs: ICommandProfileAuthConfig[] = [];
+        ImperativeConfig.instance.loadedConfig.profiles.forEach((profCfg) => {
+            if (profCfg.type === profileType && profCfg.authConfig != null) {
+                authConfigs.push(...profCfg.authConfig);
+            }
+        });
+
+        for (const authConfig of authConfigs) {
+            const authHandler: ICommandHandlerRequire = require(authConfig.handler);
+            const authHandlerClass = new authHandler.default();
+
+            if (authHandlerClass instanceof BaseAuthHandler) {
+                const [promptParams, loginHandler] = authHandlerClass.getPromptParams();
+
+                if (profile.tokenType !== promptParams.defaultTokenType) {
+                    continue;  // Don't use auth service that has mismatched token type
+                } else if (promptParams.serviceDescription != null) {
+                    this.params.response.console.log(`Logging in to ${promptParams.serviceDescription}`);
+                }
+
+                const sessCfg: ISession = await ConnectionPropsForSessCfg.addPropsOrPrompt({}, profile as ICommandArguments,
+                    { parms: this.params, doPrompting: true, requestToken: true, ...promptParams });
+                Logger.getAppLogger().info(`Fetching ${profile.tokenType} for ${propPath}`);
+
+                try {
+                    return await loginHandler(new Session(sessCfg));
+                } catch (error) {
+                    throw new ImperativeError({
+                        msg: `Failed to fetch ${profile.tokenType} for ${propPath}`,
+                        causeErrors: error
+                    });
+                }
+            }
+        }
     }
 }