+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## `5.0.0-next.202109031503`
+
+- Enhancement: Log in to authentication service to obtain token value instead of prompting for it in `config secure` command.
+
 ## `5.0.0-next.202108181618`
 
 - Breaking: Make `fail-on-error` option true by default on `zowe plugins validate` command.
@@ -101,6 +105,10 @@ All notable changes to the Imperative package will be documented in this file.
 ## `5.0.0-next.202009251501`
 
 - Enhancement: add support for CLIs that want to run as a persistent process (daemon mode).
+
+## `4.15.1`
+
+- Bugfix: Update js-yaml to resolve a potential security issue
 
 ## `4.15.0`